Fixed minor bugs
Bug #1: cir_parser.c

A benchmark had some invisible NON-ASCII characters. This
was hadled before but the code that was bypassing the characters
wasn't at the correct position.

Bug #2: lists.c

Due to copying and pasting for loops the free_lists() function was
freeing sec_list but the iteration interval was set according to the
team2_list size and not the sec_list size.