Added sparse matrix solvers for lu and chol
Option SPARSE specifies to use sparse matrix representation.
Added Option Cases:
 .OPTIONS SPARSE     -> sparse lu
 .OPTIONS SPARSE SPD -> sparse cholesky

TODO: implement sparse bi-CG and CG methods.

BUGFIX#1:
If a component has both of its nodes grounded the mna_array assignment
would be out of bounds. We were not checking that case.
For example (R1 0 0 15)

BUGFIX#2:
Function execute_commands() was always executing even with the command
list being empty which was leading to faulty execution. An if statement
was added at the beggining of the function to handle this situation.