Altered iterative methods to support sparse matrix
CG and Bi-CG methods now support support sparse matrices.

Added scripts for ibm benchmark error checking.

Iterative methods now loop MAX(100, mna_dimension) times instead of 100
times. For bigger array dimensions, in order to have a correct results a
larger iteration bound is needed.

BUGFIX#1:
During the execute_commnads() function the vector b was being completely
rebuilt at each .DC iteration when the matrix was in sparse form.
This was indeed necessary, but only for the sparse chol and sparse lu
solvers (not for the sparse iterative solvers)

BUGFIX#2:
The compressed column vector had to be set to NULL after its free() inside
the lu/chol sparse decomposition functions. This is necessary because with the
iterative methods this vector is freed at the end (as there is no
decomposition whatsoever). Thus the if statement at the end of the
program (if vec==NULL) was always executing.